+++ src/silentRotate.lua
@@ -205,7 +205,7 @@ end
 -- Sends rotation setup to raid channel
 function SilentRotate:printRotationSetup()
 
-    if (IsInRaid() or SilentRotate.testMode) then
+    if SilentRotate:isActive() then
         SilentRotate:sendRotationMessage('--- ' .. SilentRotate.constants.printPrefix .. SilentRotate:getBroadcastHeaderText() .. ' ---', channel)
 
         if (SilentRotate.db.profile.useMultilineRotationReport) then