+++ changelog.md
@@ -2,8 +2,9 @@
 
 #### v0.9.3-beta (2021-xx-yy)
 
-- Added Ritual of Souls mode
+- Added Ritual of Souls mode (TBC)
 - Classic Era now completely ignores TBC modes, instead of just hiding them
+- Tranqshot and Loatheb require a raid, other modes work in dungeons as well
 
 #### v0.9.2-beta (2021-11-07)